chore(crypto): Remove some dead code mentioning the sled_cryptostore feature
