+++ crates/matrix-sdk-crypto/src/backups/mod.rs
@@ -463,57 +463,4 @@ mod test {
 
         backup_flow(machine).await
     }
-
-    #[async_test]
-    #[cfg(feature = "sled_cryptostore")]
-    async fn default_store_backups() -> Result<(), OlmError> {
-        use tempfile::tempdir;
-
-        let tmpdir = tempdir().expect("Can't create a temporary dir");
-        let machine =
-            OlmMachine::with_default_store(alice_id(), alice_device_id(), tmpdir.as_ref(), None)
-                .await?;
-
-        backup_flow(machine).await
-    }
-
-    #[async_test]
-    #[cfg(feature = "sled_cryptostore")]
-    async fn recovery_key_storing() -> Result<(), OlmError> {
-        use tempfile::tempdir;
-
-        let tmpdir = tempdir().expect("Can't create a temporary dir");
-        let machine = OlmMachine::with_default_store(
-            alice_id(),
-            alice_device_id(),
-            tmpdir.as_ref(),
-            Some("test"),
-        )
-        .await?;
-        let backup_machine = machine.backup_machine();
-
-        let recovery_key = RecoveryKey::new().expect("Can't create new recovery key");
-        let encoded_key = recovery_key.to_base64();
-
-        backup_machine.save_recovery_key(Some(recovery_key), Some("1".to_owned())).await?;
-
-        let loded_backup = backup_machine.get_backup_keys().await?;
-
-        assert_eq!(
-            encoded_key,
-            loded_backup
-                .recovery_key
-                .expect("The recovery key wasn't loaded from the store")
-                .to_base64(),
-            "The loaded key matches to the one we stored"
-        );
-
-        assert_eq!(
-            Some("1"),
-            loded_backup.backup_version.as_deref(),
-            "The loaded version matches to the one we stored"
-        );
-
-        Ok(())
-    }
 }
